code refactoring + add locale field to user table
